ppdiffusers/examples/consistency_distillation/lcm_trainer/requirements.txt
ppdififusers>=0.19.4
paddlenlp>=2.7.1
visualdl
